Abstract and move tryptich to components folder
